@@ -14,18 +14,20 @@
     "istanbul": "^0.4.4",
     "jshint": "^2.9.2",
     "mocha": "^3.0.1",
-    "tslint": "^3.14.0",
-    "typescript": "^1.8.10"
+    "tslint": "^4.5.1",
+    "typescript": "2.0.6",
+    "@types/node": "^7.0.5"
   },
   "scripts": {
-    "lint": "jshint --show-non-errors . && tslint -c ../../tslint.json \"{lib,.}/*.ts\"",
+    "lint": "tslint --type-check --project . -c ../../tslint.json \"{src,.}/*.ts\"",
+    "build": "tsc",
     "unittest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot test/_*_test.js",
     "alltest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot test/_*_test*.js",
     "unittest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test.js",
     "alltest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec test/_*_test*.js",
-    "ci": "npm -s run lint && npm -s run alltest-min && npm -s run check-cover",
-    "test": "npm -s run lint && npm -s run unittest",
-    "check-cover": "istanbul check-coverage --statements 98 --branches 100 --functions 96 --lines 98"
+    "ci": "npm -s run lint && npm -s run build && npm -s run alltest-min && npm -s run check-cover",
+    "test": "npm -s run lint && npm -s run build && npm -s run unittest",
+    "check-cover": "istanbul check-coverage --statements 99 --branches 96 --functions 97 --lines 99"
   },
   "engines": {
     "node": ">= 0.10"
