@@ -2,14 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import { EventEmitter } from 'events';
+import { Message } from './lib/message';
 
-export { anHourFromNow, encodeUriComponentStrict } from './lib/authorization';
-export import ConnectionString = require('./lib/connection_string');
 export import endpoint = require('./lib/endpoint');
 export import errors = require('./lib/errors');
 export import results = require('./lib/results');
-export import Message = require('./lib/message');
-export import SharedAccessSignature = require('./lib/shared_access_signature');
+export { anHourFromNow, encodeUriComponentStrict } from './lib/authorization';
+export { ConnectionString } from './lib/connection_string';
+export { Message }
+export { SharedAccessSignature } from './lib/shared_access_signature';
 
 // Typescript only, used by other modules in azure-iot-sdk
 export interface X509 {
