@@ -1,7 +1,12 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-export function anHourFromNow(): number;
-export function encodeUriComponentStrict(str: string): string;
-export function stringToSign(resourceUri: string, expiry: string): string;
-export function hmacHash(password: string, stringToSign: string): string;
+/*! Copyright (c) Microsoft. All rights reserved.
+ *! Licensed under the MIT license. See LICENSE file in the project root for full license information.
+ */
+/**
+ * Returns the seconds elapsed since 1 January 1970 00:00:00 UTC until one
+ * hour from now.
+ * @function anHourFromNow
+ */
+export declare function anHourFromNow(): number;
+export declare function encodeUriComponentStrict(str: string): string;
+export declare function stringToSign(resourceUri: string, expiry: string): string;
+export declare function hmacHash(password: string, stringToSign: string): string;
